working on comments reducer.  Unit test cleanup
