--- src/utils/GroupMenuHelper.js
@@ -1,3 +1,0 @@
-import path from "path-extra";
-import {generateCheckPath, loadCheckData} from "./CheckDataHelper";
-
